requirements.txt
accelerate==0.31.0
arrow==1.3.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
codecarbon==2.5.0
diffusers==0.29.0
filelock==3.15.1
fsspec==2024.6.0
huggingface-hub==0.23.3
idna==3.7
importlib-metadata==7.1.0
jinja2==3.1.4
markupsafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
prometheus-client==0.20.0
protobuf==5.27.1
psutil==5.9.8
py-cpuinfo==9.0.0
pynvml==11.5.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
rapidfuzz==3.9.4
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
sympy==1.12.1
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
types-python-dateutil==2.9.0.20240316
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
zipp==3.19.2
